feat: add authentication on apis

+++ models/todo.go
@@ -1,6 +1,8 @@
 package models
 
 import (
+	"errors"
+
 	"github.com/moneychien19/todo-list-api-go/db"
 )
 
@@ -8,6 +10,7 @@ type Todo struct {
 	Id int64
 	Title string `binding:"required"`
 	Description string `binding:"required"`
+	CreatedBy string
 }
 
 func GetTodos(page, limit int64) ([]Todo, error) {
@@ -21,7 +24,7 @@ func GetTodos(page, limit int64) ([]Todo, error) {
 	var todos []Todo
 	for rows.Next() {
 		var todo Todo
-		err = rows.Scan(&todo.Id, &todo.Title, &todo.Description)
+		err = rows.Scan(&todo.Id, &todo.Title, &todo.Description, &todo.CreatedBy)
 		if err != nil {
 			return nil, err
 		}
@@ -30,26 +33,37 @@ func GetTodos(page, limit int64) ([]Todo, error) {
 	return todos, nil
 }
 
-func (todo *Todo) CreateTodo() (*Todo, error) {
-	query := "INSERT INTO todos (title, description) VALUES (?, ?)"
+func (todo *Todo) CreateTodo(userEmail string) (*Todo, error) {
+	query := "INSERT INTO todos (title, description, createdBy) VALUES (?, ?, ?)"
 	stmt, err := db.DB.Prepare(query)
 	if err != nil {
 		return nil, err
 	}
 	defer stmt.Close()
 
-	result, err := stmt.Exec(todo.Title, todo.Description)
+	result, err := stmt.Exec(todo.Title, todo.Description, userEmail)
 	if err != nil {
 		return nil, err
 	}
 
 	id, _ := result.LastInsertId()
 	todo.Id = id
+	todo.CreatedBy = userEmail
 	
 	return todo, nil
 }
 
-func (todo *Todo) UpdateTodoById(id int64) (*Todo, error) {
+func (todo *Todo) UpdateTodoById(id int64, userEmail string) (*Todo, error) {
+	var createdBy string
+	queryCheck := "SELECT createdBy FROM todos WHERE id = ?"
+	err := db.DB.QueryRow(queryCheck, id).Scan(&createdBy)
+	if err != nil {
+		return nil, err
+	}
+	if createdBy != userEmail {
+		return nil, errors.New("unauthorized to update this todo")
+	}
+
 	query := "UPDATE todos SET title = ?, description = ? WHERE id = ?"
 	stmt, err := db.DB.Prepare(query)
 	if err != nil {
@@ -62,11 +76,22 @@ func (todo *Todo) UpdateTodoById(id int64) (*Todo, error) {
 		return nil, err
 	}
 
-	todo.Id = id	
+	todo.Id = id
+	todo.CreatedBy = userEmail
 	return todo, nil
 }
 
-func DeleteTodoById(id int64) error {
+func DeleteTodoById(id int64, userEmail string) error {
+	var createdBy string
+	queryCheck := "SELECT createdBy FROM todos WHERE id = ?"
+	err := db.DB.QueryRow(queryCheck, id).Scan(&createdBy)
+	if err != nil {
+		return err
+	}
+	if createdBy != userEmail {
+		return errors.New("unauthorized to delete this todo")
+	}
+	
 	query := "DELETE FROM todos WHERE id = ?"
 	stmt, err := db.DB.Prepare(query)
 	if err != nil {
